move FullBundle params in class
